requirements.txt
boto3
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
opensearch-py>=2.4.0
python-dotenv>=1.0.0
pydantic
pyarrow
pyewts
ruff
pyright
ty
orjson>=3.9.0
rdflib>=7.0.0
requests>=2.31.0
lxml>=6.0.2
lxml-stubs
datasketch
datasets
fasttext-wheel